RetainInstance removed from news list
