+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".view.MainActivity"
+              android:configChanges="orientation|screenSize"
               android:launchMode="singleTop">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
