+++ app/src/main/res/layout/item_comment.xml
@@ -33,8 +33,8 @@
                 android:layout_width="wrap_content"
                 android:layout_toRightOf="@+id/imgProfilePic"
                 android:layout_toEndOf="@+id/imgProfilePic"
-                android:layout_alignLeft="@+id/txtTime"
-                android:layout_alignStart="@+id/txtTime"
+                android:layout_toLeftOf="@+id/txtTime"
+                android:layout_toStartOf="@+id/txtTime"
                 android:id="@+id/txtName"
                 android:textColor="@android:color/black"
                 android:textSize="14sp"
